refactor: added classes to abstract code that deals with azure sdk

+++ pypeliner/contrib/azure/batchqueue.py
@@ -1,28 +1,17 @@
 from __future__ import print_function
 
 import datetime
-import os
-import time
-import pickle
 import logging
-import yaml
+import os
+import pickle
+import time
 
 import pypeliner.execqueue.base
-
-import azure.storage.blob as azureblob
-import azure.batch.batch_service_client as batch
-import azure.batch.models as batchmodels
-
-from azure.profiles import KnownProfiles
-
-from azure.common.credentials import ServicePrincipalCredentials
-
-KnownProfiles.default.use(KnownProfiles.latest)
-
-import pypeliner.contrib.azure.helpers as azure_helpers
-
-import pypeliner.tests.jobs
 import pypeliner.helpers
+import pypeliner.tests.jobs
+import yaml
+
+from .batchclient import BatchClient
 
 
 class AzureJobQueue(object):
@@ -30,7 +19,6 @@ class AzureJobQueue(object):
     """
 
     def __init__(self, config_filename=None, **kwargs):
-        azure_helpers.set_azure_logging_filters()
         batch_account_url = os.environ['AZURE_BATCH_URL']
         client_id = os.environ['CLIENT_ID']
         tenant_id = os.environ['TENANT_ID']
@@ -41,42 +29,16 @@ class AzureJobQueue(object):
             self.config = yaml.load(f)
 
         storage_account_name = self.config['pypeliner_storage_account']
-        storage_account_key = azure_helpers.get_storage_account_key(
-            storage_account_name, client_id, secret_key,
-            tenant_id, keyvault_account)
+        storage_keys = pypeliner.helpers.GlobalState.get('azure_storage_keys', {})
+        account_key = storage_keys.get(storage_account_name)
 
-        self.logger = logging.getLogger('pypeliner.execqueue.azure_batch')
-
-        if 'run_id' in self.config:
-            self.run_id = self.config['run_id']
-        else:
-            self.run_id = azure_helpers.random_string(8)
-
-        self.logger.info('creating blob client')
-
-        self.blob_client = azureblob.BlockBlobService(
-            account_name=storage_account_name,
-            account_key=storage_account_key)
-
-        self.credentials = ServicePrincipalCredentials(client_id=client_id,
-                                                       secret=secret_key,
-                                                       tenant=tenant_id,
-                                                       resource="https://batch.core.windows.net/")
-
-        self.logger.info('creating batch client')
-
-        self.batch_client = batch.BatchServiceClient(
-            self.credentials,
-            batch_account_url
+        self.batch_client = BatchClient(
+            client_id, tenant_id, secret_key, batch_account_url,
+            self.config, keyvault_account, storage_account_key=account_key
         )
 
-        self.logger.info('creating task container')
-
-        self.container_name = self.config['storage_container_name']
-        self.blob_client.create_container(self.container_name)
-
-        self.compute_start_commands = {id: pool['compute_start_commands'] for id,pool in self.config['pools'].items()}
-        self.compute_finish_commands = {id: pool['compute_finish_commands'] for id,pool in self.config['pools'].items()}
+        self.logger = self.batch_client.logger
+        self.run_id = self.batch_client.run_id
 
         self.no_delete_pool = self.config.get('no_delete_pool', False)
         self.no_delete_job = self.config.get('no_delete_job', False)
@@ -92,9 +54,6 @@ class AzureJobQueue(object):
         self.completed_task_ids = set()
         self.running_task_ids = set()
 
-        self.active_pools = set()
-        self.active_jobs = set()
-
     debug_filenames = {
         'job stderr': 'stderr.txt',
         'job stdout': 'stdout.txt',
@@ -102,68 +61,16 @@ class AzureJobQueue(object):
         'upload out': 'fileuploadout.txt',
     }
 
-
     def __enter__(self):
         return self
 
     def __exit__(self, exc_type, exc_value, traceback):
         self.logger.info('tear down')
-
-        for pool_id in self.active_pools:
-
-            if not self.no_delete_pool:
-                self.logger.info("deleting pool {}".format(pool_id))
-                try:
-                    self.batch_client.pool.delete(pool_id)
-                except Exception as e:
-                    print(e)
-
-        for job_id in self.active_jobs:
-            if not self.no_delete_job:
-                self.logger.info("deleting job {}".format(job_id))
-                try:
-                    self.batch_client.job.delete(job_id)
-                except Exception as e:
-                    print(e)
-
-        for job_id in self.active_jobs:
-            if not self.no_delete_job:
-                azure_helpers.wait_for_job_deletion(self.batch_client, job_id, self.logger)
+        self.batch_client.delete_pools(ignore_exception=True)
+        self.batch_client.delete_jobs(ignore_exception=True, wait=True)
 
     def get_jobid(self, pool_id):
         return 'pypeliner_{}_{}'.format(pool_id, self.run_id)
-
-    def prep_pools_and_jobs(self, pool_id, job_id):
-        """
-        start a pool and job for the task if it hasnt been initialized already
-        :param str pool_id: azure batch pool id
-        :param str job_id: azure batch job id
-        """
-
-        if pool_id not in self.active_pools and \
-                not self.batch_client.pool.exists(pool_id):
-            pool_params = self.config['pools'][pool_id]
-            self.logger.info("creating pool {}".format(pool_id))
-            azure_helpers.create_pool(
-                self.batch_client,
-                self.blob_client,
-                pool_id,
-                pool_params)
-            self.active_pools.add(pool_id)
-
-        job_ids = set([job.id for job in self.batch_client.job.list()])
-        if job_id not in self.active_jobs and job_id not in job_ids:
-            self.logger.info("creating job {}".format(job_id))
-            azure_helpers.create_job(
-                self.batch_client,
-                job_id,
-                pool_id)
-            self.active_jobs.add(job_id)
-
-            for task in self.batch_client.task.list(job_id):
-                self.batch_client.task.delete(job_id, task.id)
-
-        return job_id
 
     def send(self, ctx, name, sent, temps_dir):
         """ Add a job to the queue.
@@ -175,7 +82,7 @@ class AzureJobQueue(object):
             temps_dir (str): unique path for strong job temps
 
         """
-        pool_id = azure_helpers.find_pool(self.config['pools'], ctx)
+        pool_id = self.batch_client.find_pool(ctx)
 
         job_id = self.get_jobid(pool_id)
 
@@ -183,9 +90,10 @@ class AzureJobQueue(object):
             'assigning job {} to pool {}'.format(name, pool_id)
         )
 
-        self.prep_pools_and_jobs(pool_id, job_id)
+        self.batch_client.create_pool(pool_id)
+        self.batch_client.create_job(pool_id, job_id)
 
-        task_id = azure_helpers.get_task_id(name)
+        task_id = self.batch_client.get_task_id(name)
 
         self.logger.info(
             'assigning task_id {} to job {}'.format(
@@ -209,8 +117,9 @@ class AzureJobQueue(object):
         with open(job_before_filename, 'wb') as before:
             pickle.dump(sent, before)
 
-        job_before_file = azure_helpers.create_blob_batch_resource(
-            self.blob_client, self.container_name, job_before_filename, job_before_blobname, job_before_file_path)
+        job_before_file = self.batch_client.create_blob_resource(
+            job_before_filename, job_before_blobname, job_before_file_path
+        )
 
         job_after_file_path = 'job_result.pickle'
         job_after_filename = os.path.join(temps_dir, job_after_file_path)
@@ -219,57 +128,45 @@ class AzureJobQueue(object):
             job_after_file_path)
 
         # Delete any previous job result file locally and in blob
-        azure_helpers.delete_from_container(
-            self.blob_client, self.container_name, job_after_blobname)
-        try:
+        self.batch_client.blob_client.delete_blob(
+            container_name=self.batch_client.container_name, blob_name=job_after_blobname
+        )
+        if os.path.exists(job_after_filename):
             os.remove(job_after_filename)
-        except OSError:
-            pass
 
         # Delete any previous log files locally
         for debug_type, debug_filename in self.debug_filenames.items():
             debug_filename = os.path.join(temps_dir, debug_filename)
-            try:
-                os.remove(debug_filename)
-            except OSError:
-                pass
+            if os.path.exists(debug_filename):
+               os.remove(debug_filename)
 
         # Create a bash script for running the job
         run_script_file_path = 'job.sh'
         run_script_filename = os.path.join(temps_dir, run_script_file_path)
         run_script_blobname = os.path.join(
-            self.job_blobname_prefix[name],
-            run_script_file_path)
+            self.job_blobname_prefix[name], run_script_file_path
+        )
 
-        command = azure_helpers.get_run_command(ctx)
-        command = command.format(
-            input_filename=job_before_file_path,
-            output_filename=job_after_file_path)
+        command = self.batch_client.get_run_command(
+            ctx, job_before_file_path, job_after_file_path
+        )
 
         with open(run_script_filename, 'w') as f:
             f.write('set -e\n')
             f.write('set -o pipefail\n\n')
-            f.write(self.compute_start_commands[pool_id] + '\n')
+            f.write(self.batch_client.compute_start_commands[pool_id] + '\n')
             f.write(command + '\n')
-            f.write(self.compute_finish_commands[pool_id] + '\n')
+            f.write(self.batch_client.compute_finish_commands[pool_id] + '\n')
             f.write('wait')
 
-        run_script_file = azure_helpers.create_blob_batch_resource(
-            self.blob_client, self.container_name, run_script_filename, run_script_blobname, run_script_file_path)
-
-        # Obtain a shared access signature that provides write access to the output
-        # container to which the tasks will upload their output.
-        container_sas_token = azure_helpers.get_container_sas_token(
-            self.blob_client,
-            self.container_name,
-            azureblob.BlobPermissions.CREATE | azureblob.BlobPermissions.WRITE)
+        run_script_file = self.batch_client.create_blob_resource(
+            run_script_filename, run_script_blobname, run_script_file_path)
 
         # Add the task to the job
-        azure_helpers.add_task(
-            self.batch_client, job_id, task_id,
-            job_before_file, run_script_file,
-            self.container_name, container_sas_token,
-            self.job_blobname_prefix[name], self.blob_client)
+        self.batch_client.add_task(
+            job_id, task_id, job_before_file,
+            run_script_file, self.job_blobname_prefix[name]
+        )
 
         self.running_task_ids.add(task_id)
 
@@ -314,19 +211,16 @@ class AzureJobQueue(object):
                 "Most recent transition: {}".format(
                     self.most_recent_transition_time))
             task_filter = "state eq 'completed'"
-            list_options = batchmodels.TaskListOptions(filter=task_filter)
 
-            for pool_id in self.active_pools:
-                azure_helpers.check_pool_for_failed_nodes(
-                    self.batch_client,
-                    pool_id,
-                    self.logger)
+            for pool_id in self.batch_client.active_pools:
+                self.batch_client.check_pool_for_failed_nodes(pool_id)
 
-            for job_id in self.active_jobs:
+            for job_id in self.batch_client.active_jobs:
                 tasks = list(
-                    self.batch_client.task.list(
+                    self.batch_client.get_task_list(
                         job_id,
-                        task_list_options=list_options))
+                        task_filter=task_filter)
+                )
                 self.logger.info("Received total {} tasks".format(len(tasks)))
                 for task in tasks:
                     if task.id not in self.completed_task_ids:
@@ -337,6 +231,8 @@ class AzureJobQueue(object):
 
     def _update_task_state(self, latest_transition_time=None):
         """ Query azure and update task state. """
+
+        list_max_results = 1000
 
         task_filter = "state eq 'completed'"
 
@@ -354,15 +250,12 @@ class AzureJobQueue(object):
             task_filter += " and stateTransitionTime lt DateTime'{}'".format(
                 filter_time_string)
 
-        list_max_results = 1000
-        list_options = batchmodels.TaskListOptions(
-            filter=task_filter,
-            max_results=list_max_results)
-
         tasks = []
-        for job_id in self.active_jobs:
-            tasks += list(self.batch_client.task.list(job_id,
-                                                      task_list_options=list_options))
+        for job_id in self.batch_client.active_jobs:
+            tasks += list(self.batch_client.get_task_list(
+                job_id, task_filter=task_filter, list_max_results=list_max_results
+            )
+            )
         sorted_transition_times = sorted(
             [task.state_transition_time for task in tasks])
 
@@ -427,13 +320,9 @@ class AzureJobQueue(object):
 
         # hostname on the node doesnt match id in azure ui(affinity id in API)
         # overrides the hostname collected from node with affinity id.
-        hostname = self.batch_client.task.get(
-            job_id,
-            task_id).node_info.node_id
+        hostname = self.batch_client.get_node_id_for_task(job_id, task_id)
 
-        azure_helpers.download_blobs_from_container(
-            self.blob_client,
-            self.container_name,
+        self.batch_client.download_blobs_from_container(
             temps_dir,
             blobname_prefix)
 
@@ -460,13 +349,8 @@ class AzureJobQueue(object):
         # batch keeps scheduling tasks on nodes in failed state, need to
         # explicitly disable failed nodes.
         if not received.finished or not received.started:
-            pool_id = azure_helpers.get_pool_id_from_job(self.batch_client, job_id)
-
-            azure_helpers.verify_node_status(
-                hostname,
-                pool_id,
-                self.batch_client,
-                self.logger)
+            pool_id = self.batch_client.get_pool_id_from_job(job_id)
+            self.batch_client.verify_node_status(hostname, pool_id)
 
         return received
